// packages/bpk-component-calendar/test-utils.d.ts
/*
 * Backpack - Skyscanner's Design System
 *
 * Copyright 2016 Skyscanner Ltd
 *
 * Licensed under the Apache License, Version 2.0 (the "License");
 * you may not use this file except in compliance with the License.
 * You may obtain a copy of the License at
 *
 *   http://www.apache.org/licenses/LICENSE-2.0
 *
 * Unless required by applicable law or agreed to in writing, software
 * distributed under the License is distributed on an "AS IS" BASIS,
 * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 * See the License for the specific language governing permissions and
 * limitations under the License.
 */

export declare const formatDateFull: (date: Date) => string;
export declare const formatDateFullArabic: (date: Date) => string;
export declare const formatDateFullJapanese: (date: Date) => string;
export declare const formatMonth: (date: Date) => string;
export declare const formatMonthArabic: (date: Date) => string;
export declare const formatMonthJapanese: (date: Date) => string;
export declare const weekDays: {
    name: string;
    nameAbbr: string;
    nameNarrow: string;
    index: number;
    isWeekend: boolean;
}[];
export declare const weekDaysMoreWeekend: {
    name: string;
    nameAbbr: string;
    nameNarrow: string;
    index: number;
    isWeekend: boolean;
}[];
export declare const weekDaysArabic: {
    name: string;
    nameAbbr: string;
    nameShort: string;
    nameNarrow: string;
    index: number;
    cldrKey: string;
    isWeekend: boolean;
}[];
export declare const weekDaysJapanese: {
    name: string;
    nameAbbr: string;
    nameShort: string;
    nameNarrow: string;
    index: number;
    cldrKey: string;
    isWeekend: boolean;
}[];